+++ AirCasting/Graph/AirCastingGraph.swift
@@ -1,0 +1,143 @@
+// Created by Lunar on 18/06/2021.
+//
+
+import SwiftUI
+import Charts
+
+class AirCastingGraph: UIView {
+    
+    let lineChartView = LineChartView()
+    var renderer: MultiColorGridRenderer?
+    var didMoveOrScaleGraph = false
+    var limitLines: [ChartLimitLine] = [] {
+        didSet {
+            guard oldValue != limitLines else { return }
+            updateMidnightLines(with: limitLines)
+        }
+    }
+    
+    init() {
+        super.init(frame: .zero)
+        self.addSubview(lineChartView)
+        lineChartView.delegate = self
+        try? setupGraph()
+    }
+    
+    private func setupGraph() throws {
+        
+        lineChartView.translatesAutoresizingMaskIntoConstraints = false
+        NSLayoutConstraint.activate([
+            lineChartView.leadingAnchor.constraint(equalTo: self.leadingAnchor),
+            lineChartView.trailingAnchor.constraint(equalTo: self.trailingAnchor),
+            lineChartView.topAnchor.constraint(equalTo: self.topAnchor),
+            lineChartView.bottomAnchor.constraint(equalTo: self.bottomAnchor)
+        ])
+        //set edges
+        lineChartView.minOffset = 0.0
+        
+        //remove border lines and legend
+        lineChartView.xAxis.drawLabelsEnabled = false
+        
+        lineChartView.leftAxis.drawLabelsEnabled = false
+        lineChartView.leftAxis.drawAxisLineEnabled = false
+        lineChartView.leftAxis.drawGridLinesEnabled = true
+        
+        lineChartView.rightAxis.enabled = false
+        lineChartView.rightAxis.drawLabelsEnabled = false
+        lineChartView.rightAxis.drawAxisLineEnabled = false
+        
+        lineChartView.legend.enabled = false
+        lineChartView.scaleYEnabled = false
+        
+        lineChartView.xAxis.drawLabelsEnabled = true
+        lineChartView.xAxis.labelCount = 2
+        lineChartView.extraBottomOffset = 25
+        
+        lineChartView.highlightPerTapEnabled = false
+        lineChartView.highlightPerDragEnabled = false
+        
+        lineChartView.xAxisRenderer = TimeAxisRenderer(viewPortHandler: lineChartView.viewPortHandler,
+                                                       xAxis: lineChartView.xAxis,
+                                                       transformer: lineChartView.getTransformer(forAxis: .left))
+        
+        
+        renderer = MultiColorGridRenderer(viewPortHandler: lineChartView.viewPortHandler,
+                                          yAxis: lineChartView.leftAxis,
+                                          transformer: lineChartView.getTransformer(forAxis: .left))
+        guard let renderer = renderer else {
+            throw GraphError.rendererError
+        }
+        lineChartView.leftYAxisRenderer = renderer
+    }
+    
+    private func zoomoutToThirtyMinutes(dataSet: LineChartDataSet) {
+        let thirtyMinutesMeasurementCount = 60 * 30
+        lineChartView.setVisibleXRangeMaximum(Double(thirtyMinutesMeasurementCount))
+        lineChartView.moveViewToX(dataSet.xMax)
+        //enable zoom out
+        lineChartView.setVisibleXRangeMaximum(dataSet.xMax)
+    }
+    
+    func updateWithThreshold(thresholdValues: [Float]) throws {
+        guard let renderer = renderer else {
+            throw GraphError.rendererError
+        }
+        renderer.thresholds = thresholdValues
+        
+        lineChartView.leftAxis.axisMinimum = Double(thresholdValues.first ?? 0)
+        lineChartView.leftAxis.axisMaximum = Double(thresholdValues.last ?? 200)
+        
+        lineChartView.data = lineChartView.data
+        lineChartView.setNeedsDisplay()
+    }
+    
+    func updateWithEntries(entries: [ChartDataEntry], isAutozoomEnabled: Bool) {
+        let dataSet = LineChartDataSet(entries: entries)
+        let data = LineChartData(dataSet: dataSet)
+        lineChartView.data = data
+        
+        //format data labels
+        data.setDrawValues(false)
+        
+        // Line styling
+        dataSet.drawCirclesEnabled = false
+        dataSet.setColor(UIColor(.white))
+        dataSet.mode = .linear
+        dataSet.lineWidth = 4
+        
+        if !didMoveOrScaleGraph && isAutozoomEnabled {
+            zoomoutToThirtyMinutes(dataSet: dataSet)
+        }
+    }
+    
+    private func updateMidnightLines(with limitLines: [ChartLimitLine]) {
+        lineChartView.xAxis.removeAllLimitLines()
+        for line in limitLines {
+            lineChartView.xAxis.addLimitLine(line)
+        }
+    }
+    
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
+
+extension AirCastingGraph: ChartViewDelegate {
+    
+    // Callbacks when the chart is scaled / zoomed via pinch zoom gesture.
+    @objc func chartScaled(_ chartView: ChartViewBase, scaleX: CGFloat, scaleY: CGFloat) {
+        didMoveOrScaleGraph = true
+    }
+    // Callbacks when the chart is moved / translated via drag gesture.
+    @objc func chartTranslated(_ chartView: ChartViewBase, dX: CGFloat, dY: CGFloat) {
+        didMoveOrScaleGraph = true
+    }
+}
+
+#if DEBUG
+struct AirCastingGraph_Previews: PreviewProvider {
+    static var previews: some View {
+        Graph(stream: .mock, thresholds: .mock, isAutozoomEnabled: true)
+    }
+}
+#endif
